cancel booked完成，但不能刷新页面

+++ src/main/java/com/springboot/Captioner/service/BookingServiceImp.java
@@ -63,7 +63,13 @@ public class BookingServiceImp implements BookingService { // 确保实现了接
     public Booking cancelPlay(String userEmail, String playTitle) {
         // 根据用户邮箱地址和剧目标题查询预订记录
          Booking booking = bookingRepository.findByUserEmailAndPlayTitle(userEmail, playTitle);
-        bookingRepository.delete(booking);
+        System.out.println("Cancel play method called with userEmail: " + userEmail + " and playTitle: " + playTitle);
+        if (booking != null) {
+            System.out.println("Found booking for userEmail: " + userEmail + " and playTitle: " + playTitle);
+            bookingRepository.delete(booking);
+        }else {
+            System.out.println("No booking found for userEmail: " + userEmail + " and playTitle: " + playTitle);
+        }
         return booking;
     }
